refactor: altera estrutura do router
