+++ front-end/src/pages/PageHome/index.js
@@ -1,0 +1,15 @@
+import { Header } from "../../components/Header";
+import { Hero } from "../../components/Hero";
+import { BecomeMentor } from "../../components/BecomeMentor";
+import { Benefits } from "../../components/Benefits";
+
+export const PageHome = () => {
+  return (
+    <>
+      <Header />
+      <Hero />
+      <BecomeMentor />
+      <Benefits />
+    </>
+  );
+};
