+++ front-end/src/pages/PageLogin/index.js
@@ -1,0 +1,3 @@
+export const PageLogin = () => {
+  return <h1>Login Page</h1>;
+};
