+++ front-end/src/pages/PageMentor/index.js
@@ -1,0 +1,3 @@
+export const PageMentor = () => {
+  return <h1>Mentor Page</h1>;
+};
