+++ front-end/src/pages/PageSchedule/index.js
@@ -1,0 +1,3 @@
+export const PageSchedule = () => {
+  return <h1>Schedule Page</h1>;
+};
